--- src/components/shared/dummy/Dummy.tsx
@@ -1,3 +1,0 @@
-export default function Dummy() {
-  return <h1>Dummy</h1>;
-}
