+++ src/pages/index.tsx
@@ -1,10 +1,10 @@
-import Tag from '@/components/Tag';
+import Tag from '@/components/shared/Tag';
 
 export default function Home() {
   return (
     <div className="mx-5 flex space-x-3">
-      <Tag Variant="Search" Hour={9} />
-      <Tag Variant="Detail" Hour={2} />
+      <Tag Type="default" Hour={9} />
+      <Tag Type="Detail" Hour={2} />
     </div>
   );
 }
